% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HohgantR-package.R
\docType{package}
\name{HohgantR-package}
\alias{HohgantR}
\alias{HohgantR-package}
\title{HohgantR: Bernhard Bieri's personal package with a set of miscellaneous functions.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package is a personal toolbox of functions that have no place in other packages I have contributed to or developed.
}
\author{
\strong{Maintainer}: Bernhard Bieri \email{bernhard.bieri1@gmail.com} (\href{https://orcid.org/0000-0001-5943-9059}{ORCID})

}
\keyword{internal}
