% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{indicators_long}
\alias{indicators_long}
\title{Synthetic example dataset of indicators}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(indicators_long)
}
\description{
Synthetic example dataset of indicators used to power a few examples in
the package.
}
\examples{
data(indicators_long)
dplyr::glimpse(indicators_long)
}
\keyword{datasets}
